feat: nuclei committere

+++ payment-service/src/main/java/com/paymentservice/application/service/params/PayParams.java
@@ -2,8 +2,7 @@ package com.paymentservice.application.service.params;
 
 import com.paymentservice.domain.vo.PaymentMethod;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
-public record PayParams(UUID rideId, UUID passengerId, BigDecimal amountPaid, BigDecimal totalToPay, PaymentMethod method) {
+public record PayParams(UUID rideId, UUID passengerId, double amountPaid, double totalToPay, PaymentMethod method) {
 }
