+++ payment-service/src/main/java/com/paymentservice/presentation/controller/requests/PayRequest.java
@@ -2,8 +2,7 @@ package com.paymentservice.presentation.controller.requests;
 
 import com.paymentservice.domain.vo.PaymentMethod;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
-public record PayRequest(UUID passengerId, BigDecimal amountPaid, BigDecimal totalToPay, PaymentMethod method) {
+public record PayRequest(UUID passengerId, double amountPaid, double totalToPay, PaymentMethod method) {
 }
