+++ ride-service/src/main/java/com/rideservice/application/events/PaymentSuccessEvent.java
@@ -1,7 +1,6 @@
 package com.rideservice.application.events;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
-public record PaymentSuccessEvent(UUID rideId, BigDecimal change) {
+public record PaymentSuccessEvent(UUID rideId, double change) {
 }
