+++ ride-service/src/main/java/com/rideservice/application/service/events/RideRequestedEvent.java
@@ -2,13 +2,12 @@ package com.rideservice.application.service.events;
 
 import com.rideservice.domain.vo.location.Location;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 public record RideRequestedEvent(
-     UUID rideId,
-     UUID passengerId,
+        UUID  rideId,
+        UUID  passengerId,
      Location origin,
      Location destination,
-     BigDecimal amount){
+     double amount){
 }
