+++ ride-service/src/main/java/com/rideservice/infrastructure/repository/ride/vo/PriceEntity.java
@@ -3,12 +3,10 @@ package com.rideservice.infrastructure.repository.ride.vo;
 import jakarta.persistence.Embeddable;
 import jakarta.persistence.Embedded;
 
-import java.math.BigDecimal;
-
 @Embeddable
 public class PriceEntity {
 
-    private BigDecimal amount;
+    private double amount;
 
     @Embedded
     private LocationEntity origin;
@@ -19,13 +17,13 @@ public class PriceEntity {
     protected PriceEntity() {
     }
 
-    public PriceEntity(BigDecimal amount, LocationEntity origin, LocationEntity destination) {
+    public PriceEntity(double amount, LocationEntity origin, LocationEntity destination) {
         this.amount = amount;
         this.origin = origin;
         this.destination = destination;
     }
 
-    public BigDecimal getAmount() {
+    public double getAmount() {
         return amount;
     }
